refactor:  Added support for responsive images on restaurant info page
Created a copy of the restaurant images at 50% to accommodate  1x devices.  Also reduced file those file sizes by 30%.   Updated the json data with the new filename and duplicated the image helper function so it would fetch the smaller image. Need to review that last bit. Not super happy with duplicating code.