feat:  Added service worker and offline cache response
Created a service worker to cache basic support files so as long as the page loaded once, there would be something usable. As the user browses the site the cache extends so any page they previously viewed can be accessed offline.  Added an offline html file that the user is redirected to it if they are offline and try to load content that isn't cached.