Fixes:  Map and Breadcrumb label and role corrections.
Map was not assigned an aria-label or role.  Based on https://www.w3.org/WAI/PF/aria/roles the correct role would be application. Map is, "the element that encompasses the entire (map) application."

Breadcrumb was not assigned an aria-label or role.   Based on design pattern at https://www.w3.org/TR/wai-aria-practices/examples/breadcrumb/index.html it was assigned a label of Breadcrumb.  The navigation role would be appropriate because it is a collection of navigational elements. https://www.w3.org/WAI/PF/aria/roles#navigation